requirements.txt
streamlit
newspaper3k
tldextract
transformers
torch
scikit-learn
